+++ src/strerror.c
@@ -95,6 +95,9 @@ const char *lcb_strerror(lcb_t instance, lcb_error_t error)
     case LCB_PLUGIN_VERSION_MISMATCH:
         return "The plugin used for IO operations cannot be loaded due to"
                " a version mismatch";
+    case LCB_INVALID_HOST_FORMAT:
+        return "One of the hostnames specified use invalid characters"
+            " or an unsupported format";
     default:
         return "Unknown error.. are you sure libcouchbase gave you that?";
     }